Collapse code into single action.
