+++ crates/platform-node/src/actions/install_deps.rs
@@ -1,11 +1,14 @@
-use crate::deps::install_node_modules;
 use moon_action::{Action, ActionContext, ActionStatus};
 use moon_config::NodePackageManager;
 use moon_contract::SupportedPlatform;
-use moon_lang_node::{package::PackageJson, NPM};
-use moon_logger::{color, debug};
+use moon_error::map_io_to_fs_error;
+use moon_error::MoonError;
+use moon_lang::has_vendor_installed_dependencies;
+use moon_lang_node::{package::PackageJson, NODE, NPM};
+use moon_logger::{color, debug, warn};
+use moon_terminal::{label_checkpoint, Checkpoint};
 use moon_toolchain::tools::node::NodeTool;
-use moon_utils::fs;
+use moon_utils::{fs, is_ci, is_offline};
 use moon_workspace::{Workspace, WorkspaceError};
 use std::sync::Arc;
 use tokio::sync::RwLock;
@@ -67,16 +70,7 @@ fn add_engines_constraint(node: &NodeTool, package_json: &mut PackageJson) -> bo
     false
 }
 
-pub async fn install_deps(
-    _action: &mut Action,
-    context: &ActionContext,
-    workspace: Arc<RwLock<Workspace>>,
-    platform: &SupportedPlatform,
-) -> Result<ActionStatus, WorkspaceError> {
-    let workspace = workspace.read().await;
-    let node = workspace.toolchain.node.get()?;
-    let pm = node.get_package_manager();
-
+async fn sync_workspace(workspace: &Workspace, node: &NodeTool) -> Result<(), MoonError> {
     // Sync values to root `package.json`
     PackageJson::sync(&workspace.root, |package_json| {
         add_package_manager(node, package_json);
@@ -99,22 +93,120 @@ pub async fn install_deps(
         );
     }
 
-    // Check if the lockfile or manifest has been modified anywhere in the workspace
+    Ok(())
+}
+
+pub async fn install_deps(
+    _action: &mut Action,
+    context: &ActionContext,
+    workspace: Arc<RwLock<Workspace>>,
+    platform: &SupportedPlatform,
+    project_id: Option<&str>,
+) -> Result<ActionStatus, WorkspaceError> {
+    let workspace = workspace.read().await;
+    let node = workspace.toolchain.node.get_from_platform(platform)?;
+    let pm = node.get_package_manager();
     let lock_filename = pm.get_lock_filename();
     let manifest_filename = pm.get_manifest_filename();
-    let has_modified_files = context
-        .touched_files
-        .iter()
-        .any(|f| f.ends_with(&manifest_filename) || f.ends_with(&lock_filename));
 
-    install_node_modules(
-        &workspace,
-        platform,
-        node,
-        &workspace.root,
-        None,
-        LOG_TARGET,
-        has_modified_files,
-    )
-    .await
+    // Determine the working directory and whether lockfiles and manifests have been modified
+    let working_dir;
+    let has_modified_files;
+
+    if let Some(project_id) = project_id {
+        let project = workspace.projects.load(project_id)?;
+
+        working_dir = project.root;
+        has_modified_files = context
+            .touched_files
+            .contains(&working_dir.join(&lock_filename))
+            || context
+                .touched_files
+                .contains(&working_dir.join(&manifest_filename));
+    } else {
+        working_dir = workspace.root.clone();
+        has_modified_files = context
+            .touched_files
+            .iter()
+            .any(|f| f.ends_with(&lock_filename) || f.ends_with(&manifest_filename));
+
+        // When installing deps in the workspace root, also sync applicable settings
+        sync_workspace(&workspace, node).await?;
+    }
+
+    // Install dependencies in the current project or workspace
+    let lock_filepath = working_dir.join(&lock_filename);
+    let mut last_modified = 0;
+    let mut cache = workspace
+        .cache
+        .cache_deps_state(platform, project_id)
+        .await?;
+
+    if lock_filepath.exists() {
+        last_modified = cache.to_millis(
+            fs::metadata(&lock_filepath)
+                .await?
+                .modified()
+                .map_err(|e| map_io_to_fs_error(e, lock_filepath.clone()))?,
+        );
+    }
+
+    // Install deps if the lockfile has been modified since the last time they were installed!
+    if has_modified_files || last_modified == 0 || last_modified > cache.item.last_install_time {
+        debug!(
+            target: LOG_TARGET,
+            "Installing {} dependencies in {}",
+            platform.label(),
+            color::path(&working_dir)
+        );
+
+        // When in CI, we can avoid installing dependencies because
+        // we can assume they've already been installed before moon runs!
+        if is_ci() && has_vendor_installed_dependencies(&working_dir, &NODE) {
+            warn!(
+                target: LOG_TARGET,
+                "In a CI environment and dependencies already exist, skipping install"
+            );
+
+            return Ok(ActionStatus::Skipped);
+        }
+
+        if is_offline() {
+            warn!(
+                target: LOG_TARGET,
+                "No internet connection, assuming offline and skipping install"
+            );
+
+            return Ok(ActionStatus::Skipped);
+        }
+
+        let install_command = match node.config.package_manager {
+            NodePackageManager::Npm => "npm install",
+            NodePackageManager::Pnpm => "pnpm install",
+            NodePackageManager::Yarn => "yarn install",
+        };
+
+        println!("{}", label_checkpoint(install_command, Checkpoint::Pass));
+
+        pm.install_dependencies(node, &working_dir).await?;
+
+        if !is_ci() && node.config.dedupe_on_lockfile_change {
+            debug!(target: LOG_TARGET, "Dedupeing dependencies");
+
+            pm.dedupe_dependencies(node, &working_dir).await?;
+        }
+
+        // Update the cache with the timestamp
+        cache.item.last_install_time = cache.now_millis();
+        cache.save().await?;
+
+        return Ok(ActionStatus::Passed);
+    }
+
+    debug!(
+        target: LOG_TARGET,
+        "Lockfile has not changed since last install, skipping Node.js dependencies",
+    );
+
+    Ok(ActionStatus::Skipped)
 }
