+++ crates/platform-node/src/actions/mod.rs
@@ -1,9 +1,7 @@
 mod install_deps;
-mod install_project_deps;
 mod run_target;
 mod sync_project;
 
 pub use install_deps::install_deps;
-pub use install_project_deps::install_project_deps;
 pub use run_target::*;
 pub use sync_project::sync_project;
