+++ crates/platform-node/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod actions;
-mod deps;
 mod hasher;
 pub mod task;
